% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr_subsampling.R
\name{subsample_countmean}
\alias{subsample_countmean}
\title{Subsampling SSR Data function}
\usage{
subsample_countmean(filepath, pct = 0.1, times = 20,
  outputname = "sample")
}
\arguments{
\item{pct}{Percentage of data to subsample}

\item{times}{Number of run}

\item{outputname}{Prefix of output files}
}
\value{
4 csv files: 1) full_locustable: A locustable of the full dataset; 2) subsample_locustable: locustable of all subsample dataset; 3) subsample_list; 4)subsample_locusallelmeansd: allel mean and sd of all subsample dataset
Genind object of the whole dataset
list object of all subsample data
Matrix descriping meand and sd of the subsample dataset
}
\description{
This function subsample a portion of ssr data and calculate the mean and sd of the subsampled set
}
\examples{
subsample_countmean("ssr_input.csv",pct=0.2,times=20,outputname="speciesA")
}
\keyword{ssr}
\keyword{subsample}
